Updated with:
Sun 26 Feb 2023 10:09:14 AM AKST

Temporarily moved the Fac-Ette GemMaster faceting machine to the
bed in order to bring the Lifetime white plastic table it was
on up to the East side of the living room in front of the window.

I do want to get faceting.  Getting the  mFac-Ette GemMaster machine out of
the downstairs bedroom is key to doing so.