Added Transfer section with:
Transfer
Once the pavillion of a gemstone is cut the dop was be transferred
from crown to pavillion in order for the crown to be cut. In my
faceting class, with Max Bradley in Emerald, Australia
(FIXME: year for this?), he had us cut the pavillion first, with
crown attached using dop wax (brown dop wax?). Then pavillion was
then dopped using epoxy. The stone with 2 dops was then put in a
freezer for a few hours. The double-dopped stone was removed and
the waxed dop stick tapped. It just popped off the crown that
still needed to be faceted. Once the crown was cut the dop
attached to the pavillion was removed by soaking in alcohol.
This dissolved the epoxy holding on the pavillion dop.